requirements.txt
ipympl==0.8.2
jupyterlab==3.0.15
matplotlib==3.3.2
numpy==1.21.4
pandas==1.3.4
python-magic==0.4.18
scikit-learn==0.24.2
tqdm==4.62.3
xgboost==1.3.3
uncertainties==3.1.6